# api/swagger/swagger.yaml
swagger: '2.0'
info:
  description: The Mutably REST API
  version: 1.0.0
  title: Mutably
  contact:
    email: marcusposey@gmail.com
host: 'localhost:8080'
basePath: /api/v1
tags:
  - name: languages
    description: Languages implemented by the API
  - name: words
    description: Words from all supported languages
  - name: users
    description: User data
  - name: tokens
    description: Resource for retrieving JSON Web Tokens
schemes:
  - http
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
  Basic:
    type: apiKey
    name: Authorization
    in: header
paths:
  /languages:
    get:
      tags:
        - languages
      summary: Retrieves all supported languages
      produces:
        - application/json
      parameters: []
      responses:
        '200':
          description: success
          schema:
            type: array
            items:
              $ref: '#/definitions/Language'
        '404':
          description: no languages exist
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/languages/{id}':
    get:
      tags:
        - languages
      summary: Retrieves a specific language
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: The id of the language
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: success
          schema:
            $ref: '#/definitions/Language'
        '404':
          description: language not found
          schema:
            $ref: '#/definitions/ErrorResponse'
  /words:
    get:
      tags:
        - words
      summary: Retrieves all words
      produces:
        - application/json
      parameters: []
      responses:
        '200':
          description: success
          schema:
            type: array
            items:
              $ref: '#/definitions/Word'
        '404':
          description: no words exist
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/words{id}':
    get:
      tags:
        - words
      summary: Retrieves a specific word
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: The id of the word
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: success
          schema:
            $ref: '#/definitions/Word'
        '404':
          description: word not found
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/words/{word}/inflections':
    get:
      tags:
        - words
      summary: Retrieves an inflection table associated with the word
      produces:
        - application/json
      parameters:
        - name: word
          in: path
          required: true
          type: string
      responses:
        '200':
          description: success
          schema:
            $ref: '#/definitions/ConjugationTable'
        '404':
          description: word has no inflections
          schema:
            $ref: '#/definitions/ErrorResponse'
  /users:
    get:
      tags:
        - users
      summary: Retrieves all users
      description: >-
        This resource requires a JWT in the Authorization header. That token
        must belong to a user with the administrator role.
      produces:
        - application/json
      parameters: []
      security:
        - Bearer: []
      responses:
        '200':
          description: success
          schema:
            type: array
            items:
              $ref: '#/definitions/User'
        '401':
          description: not authorized to access resource
          schema:
            $ref: '#/definitions/ErrorResponse'
        '404':
          description: no users exist
          schema:
            $ref: '#/definitions/ErrorResponse'
    post:
      tags:
        - users
      summary: Creates a new user
      description: >-
        The client should provide a username and password as a Basic
        Authorization header. They should be joined by a colon like so -
        aUser:pwd123
      consumes:
        - application/json
      produces:
        - application/json
      security:
        - Basic: []
      responses:
        '201':
          description: user created
          schema:
            $ref: '#/definitions/TokenResponse'
        '400':
          description: failed to create user
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/users/{id}':
    get:
      tags:
        - users
      summary: Retrieves a specific user
      description: >-
        This resource requires a JWT in the Authorization header. The token must
        belong to either the requested user or one with the administrator role.
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: The id of the user
          required: true
          type: integer
          format: int64
      security:
        - Bearer: []
      responses:
        '200':
          description: success
          schema:
            $ref: '#/definitions/User'
        '401':
          description: not authorized to access resource
          schema:
            $ref: '#/definitions/ErrorResponse'
        '404':
          description: user not found
          schema:
            $ref: '#/definitions/ErrorResponse'
  /tokens:
    get:
      tags:
        - tokens
      summary: Get a new JWT
      description: >-
        The client should provide a username and password as a Basic
        Authorization header. They should be joined by a colon like so -
        aUser:pwd123
      produces:
        - application/json
      parameters: []
      security:
        - Bearer: []
      responses:
        '200':
          description: success
          schema:
            $ref: '#/definitions/TokenResponse'
        '401':
          description: invalid user credentials
          schema:
            $ref: '#/definitions/ErrorResponse'
definitions:
  User:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      role:
        type: integer
        format: int64
      target-language:
        type: integer
        format: int64
      creation-tz:
        type: string
        format: date-time
  Language:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
      tag:
        type: string
  Word:
    type: object
    properties:
      id:
        type: integer
        format: int64
      text:
        type: string
      language:
        type: integer
        format: int64
  ConjugationTable:
    description: Present- and past-tense inflections of a verb
    type: object
    properties:
      Infinitive:
        type: string
      Present:
        $ref: '#/definitions/VerbTense'
      Past:
        $ref: '#/definitions/VerbTense'
  VerbTense:
    description: The inflections of an infinitive verb in a single tense
    type: object
    properties:
      First:
        type: array
        items:
          type: string
      Second:
        type: array
        items:
          type: string
      Third:
        type: array
        items:
          type: string
      Plural:
        type: array
        items:
          type: string
  ErrorResponse:
    description: A detailed error message
    type: object
    properties:
      error:
        type: string
  TokenResponse:
    description: A JWT
    type: object
    properties:
      token:
        type: string